pyproject.toml
[tool.poetry]
name = "wb-zero-supply"
version = "0.1.0"
description = ""
authors = ["rosoporto <rosoporto@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
python-telegram-bot = "^13.15"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
load = "wb_zero_supply.main:main"
check_domen = "wb_zero_supply.get_stock_wb_from_domen:main"
check_wb_api = "wb_zero_supply.get_stock_wb_from_api:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
